requirements.txt
accelerate==0.17.1
albumentations==1.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
einops==0.6.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
imageio==2.37.0
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kornia==0.6.10
lightning-utilities==0.15.2
markupsafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==1.4.4
pillow==11.3.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywavelets==1.4.1
pyyaml==6.0.2
qudida==0.0.4
scienceplots==2.1.0
scikit-image==0.19.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.8.1
torchvision==0.20.1+cu121
tqdm==4.64.1
typing-extensions==4.15.0
yacs==0.1.8
